+++ history.md
@@ -10,6 +10,9 @@
 - Import的complex属性适配，接收一个复杂对象实现，具体的名称和URL解析考虑单独参数或者额外包装
 ### Doing
 
+### 4.3.16
+- 修正ButtonView的upload传参错误的BUG
+
 ### 4.3.15
 - 优化undefined校验